+++ src/i18n-l10n-editor/services/translations/ije-amazon-translator.ts
@@ -12,21 +12,21 @@ export class IJEAmazonTranslator implements IJETranslation {
     results: { [language: string]: string };
     _manager: IJEManager;
     async translate(text: string, translation: IJEDataTranslation, language: string, languages: string[]): Promise<{ [language: string]: string }> {
-        const apiKey = IJEConfiguration.TRANSLATION_SERVICE_API_KEY;
+        const apiKey = IJEConfiguration.TRANSLATION_SERVICE_AMAZON_KEY || IJEConfiguration.TRANSLATION_SERVICE_API_KEY;
         if (apiKey && apiKey.length === 0) {
             vscode.window.showErrorMessage('Your Amazon API Key is blank. please update setting i18nJsonEditor.translationServiceApiKey');
 
             return { [language]: text };
         }
 
-        const apiSecret = IJEConfiguration.TRANSLATION_SERVICE_API_SECRET;
+        const apiSecret = IJEConfiguration.TRANSLATION_SERVICE_AMAZON_SECRET || IJEConfiguration.TRANSLATION_SERVICE_API_SECRET;
         if (apiSecret && apiSecret.length === 0) {
             vscode.window.showErrorMessage('Your Amazon API Secret is blank. please update setting i18nJsonEditor.translationServiceApiKey');
 
             return { [language]: text };
         }
 
-        const apiRegion = IJEConfiguration.TRANSLATION_SERVICE_API_REGION;
+        const apiRegion = IJEConfiguration.TRANSLATION_SERVICE_AMAZON_REGION || IJEConfiguration.TRANSLATION_SERVICE_API_REGION;
         if (apiRegion && apiRegion.length === 0) {
             vscode.window.showErrorMessage('Your Amazon API Region is blank. please update setting i18nJsonEditor.translationServiceApiRegion');